requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
pygame>=2.5.0
picamera2>=0.3.12
numpy>=1.24.0
RPi.GPIO>=0.7.1
requests>=2.31.0
sqlalchemy>=2.0.0
pandas>=2.0.0
pymysql>=1.1.0